feat: integrate jest-environment-emit
