+++ package.json
@@ -103,7 +103,7 @@
   "dependencies": {
     "@noomorph/allure-js-commons": "^2.3.0",
     "ci-info": "^3.8.0",
-    "jest-metadata": "^1.2.1",
+    "jest-metadata": "1.3.0-beta.1",
     "lodash.snakecase": "^4.1.1",
     "node-fetch": "^2.6.7",
     "pkg-up": "^3.1.0",
